Missing textures and yet another stuff fix (#24)
* Headgear textures and stat fix

Missing headgear textures and straw hat stat fix to make it not connected to stuffed items in any way since it's freaking not stuffed. Suspecting that straw hat textures are fucked either way cause it's case sensitive but Github is apparently not.

* Two-step straw hat fix

Gotta delete and re-add the texture cause of the fucking wrong case letter. Deleting here.

* Two-step Straw Hat fix

Deleting and re-adding texture. Adding here.